refactor: contexts 폴더 분리
